server/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
PyJWT==2.11.0
passlib[bcrypt]==1.7.4
aiosqlite==0.20.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
slowapi==0.1.9  # Rate limiting for FastAPI
